Fix shift and add program
My earliest attempt at verification was a program that did the following.

It would take the contents of a register, and decrement the register until it was zero.
Then, it would set the register to 1.
Then, it would shift the register right until it equalled 8.
Then, it would terminate.

This program demonstrates self looping and has a large state space (256 possible paths), so it made a good sample program for Klee.

This program has been updated to use the most modern versions of my verification model.