Allow "NONE" as a CPU register definition.
The C register number is allowed to be NONE, which is to be interpreted as no register to clock C into.
The computation is still assumed to have occurred, since computations may modify status bits.
This form of computation is a commonly used feature in microcode to check if a a register is even or odd.